# Conflicts:
#	Assets/PurrNet/Editor/AssetScanning/UnifiedAssetPostprocessor.cs
#	Assets/PurrNet/package.json
#	README.md

+++ Assets/PurrNet/Runtime/Components/NetworkPrefabs/AddressableNetworkPrefabs.cs
@@ -143,7 +143,7 @@ namespace PurrNet
         public bool TryGetPersistentId(GameObject prefab, out string persistentId)
         {
             if (TryGetPrefabData(prefab, out var prefabData))
-                return TryGetPersistentId(prefabData.prefabId, out persistentId);
+                return TryGetPersistentId((int)prefabData.prefabId, out persistentId);
 
             persistentId = null;
             return false;
